+++ pkg/lbaas/server/api.go
@@ -2,7 +2,7 @@ package server
 
 import (
 	"context"
-	"github.com/anexia-it/go-anxcloud/pkg/lbaas/pagination"
+	"github.com/anexia-it/go-anxcloud/pkg/pagination"
 
 	"github.com/anexia-it/go-anxcloud/pkg/client"
 )
